scripts/python/requirements.txt
requests==2.31.0
anthropic==0.45.0
pysrt==1.1.2
python-dotenv==1.0.0
tqdm==4.66.1
retry==0.9.2